src/requirements.txt
click~=8.1.3
fastapi~=0.95.1
uvicorn~=0.21.1
jinja2~=3.1.2
python-magic~=0.4.27
humanize~=4.6.0